Added files via upload
You are required to implement a Painting manager system in C++. This system keeps track of paintings

in an art gallery. Each painting contains the following fields:

1. ID—This is a unique (integer) identifier for each painting

2. Title—This is the title of the painting.

3. Artist name—The first name of the artist who painted the painting.

4. Picture—A file containing a picture of the painting in an appropriate format

The last item above is a jpeg-formatted sequence meant to display an image of a painting on the

user’s display. For the time being, you do not have to worry about creating, storing or displaying the

image; you may assume that this functionality will be implemented by someone else. However, you must

make provisions for the fact that images will be relatively large, making each painting instance expensive

to store in memory.

To handle the size of each painting instance, your manager (Mr. Oogawh Booey) has decided that

you should